+++ CryptokiMan/digest_dlg.h
@@ -52,7 +52,7 @@ private slots:
     void runFileDigestThread();
     void startTask();
     void onTaskFinished();
-    void onTaskUpdate( int nUpdate );
+    void onTaskUpdate( qint64 nUpdate );
 
 private:
     void initialize();
