+++ CryptokiMan/digest_thread.cpp
@@ -36,7 +36,7 @@ void DigestThread::run()
 
     int nRead = 0;
     int nPartSize = manApplet->settingsMgr()->fileReadSize();
-    int nReadSize = 0;
+    qint64 nReadSize = 0;
     int nLeft = 0;
     int nOffset = 0;
