+++ CryptokiMan/encrypt_dlg.h
@@ -61,7 +61,7 @@ private slots:
     void runFileEncryptThread();
     void startTask();
     void onTaskFinished();
-    void onTaskUpdate( int nUpdate );
+    void onTaskUpdate( qint64 nUpdate );
 
 private:
     void initialize();
